+++ src/Xeora.Web.Service.DSS/IManager.cs
@@ -2,6 +2,6 @@
 {
     public interface IManager
     {
-        void Reserve(string uniqueId, int reservationTimeout, out Basics.Dss.IDss reservationObject);
+        void Reserve(string uniqueId, short reservationTimeout, out Basics.Dss.IDss reservationObject);
     }
 }
